Tavily_ev-automation-main/evAutomationUpdated/pyproject.toml
[project]
name = "ev-llm-comparison"
version = "0.1.0"
description = "Production Excel-first pipeline for LLM comparison with RAG and RAGAS."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "datasets>=2.20.0",
    "google-genai>=1.0.0",
    "google-generativeai>=0.8.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-huggingface>=0.1.2",
    "langchain-google-genai>=2.0.0",
    "langchain-ollama>=0.2.0",
    "ollama>=0.6.0",
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    "pypdf>=5.0.0",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.12.0",
    "ragas>=0.2.0",
    "sentence-transformers>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ev_llm_compare"]
